% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileGraphs-package.R
\docType{package}
\name{SoilProfileGraphs-package}
\alias{SoilProfileGraphs}
\alias{SoilProfileGraphs-package}
\title{SoilProfileGraphs: Tools for Soil Profile Plots}
\description{
Provides extensions for 'aqp' and 'ggplot2' for creation of soil profile-related graphics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brownag/SoilProfileGraphs}
  \item \url{http://humus.rocks/SoilProfileGraphs/}
  \item Report bugs at \url{https://github.com/brownag/SoilProfileGraphs/issues}
}

}
\author{
\strong{Maintainer}: Andrew Brown \email{brown.andrewg@gmail.com}

}
\keyword{internal}
