% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spg.R
\name{spg}
\alias{spg}
\title{spg}
\usage{
spg(
  spc,
  filename,
  prop,
  destfile = NULL,
  depth_min = 0,
  depth_max = 200,
  prop_min = NULL,
  prop_max = NULL,
  spline = FALSE,
  midpoint = FALSE,
  xlab = prop,
  ylab = "Depth, cm",
  no_data_value = NULL,
  panel_ontop = TRUE,
  prop_linetype = "longdash",
  prop_linewidth = 1,
  prop_color = "yellow",
  prop_breaks = 6
)
}
\arguments{
\item{spc}{SoilProfileCollection}

\item{filename}{filename}

\item{prop}{character. Property column name}

\item{destfile}{Optional: output filename}

\item{depth_min}{Minimum image depth (cm)}

\item{depth_max}{Maximum image depth (cm)}

\item{prop_min}{Property minimum value}

\item{prop_max}{Property maximum value}

\item{spline}{Default: \code{FALSE}}

\item{midpoint}{Default: \code{FALSE} If \code{spline=FALSE} use midpoints of horizons?}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{no_data_value}{Replacement value for \code{NA} in \code{prop}}

\item{panel_ontop}{Default: \code{TRUE}}

\item{prop_linetype}{Default: \code{longdash}}

\item{prop_linewidth}{Default: \code{1}}

\item{prop_color}{Default: \code{"yellow"}}

\item{prop_breaks}{Default: \code{6}}
}
\value{
ggplot plot object; also shown on graphics pane as side-effect.
}
\description{
soil profile property graphs with 'calibrated' pit photo backdrop.
}
\examples{
\dontshow{if (!inherits(requireNamespace("soilDB", quietly=TRUE), 'try-error')) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
\donttest{
library(soilDB)

spc <- readRDS(system.file("extdata", "aecpedons.rds",
                   package = "SoilProfileGraphs"))[4, ]
lab <- soilDB::fetchLDM(spc$pedlabsampnum)

spg(
  system.file("extdata",
              "sueredo.png",
              package = "SoilProfileGraphs"),
  spc = lab,
  prop = "sand_total",
  depth_max = 185,
  prop_min = 65,
  prop_max = 95
)
}
}
